Refactor: 디렉토리 구조 변경 Resolve #19
utilFunction.js 삭제 후 기능별로 분리하여 관리

- 입력창 하단 메시지 관련 기능 -> /js/message.js
- 약관 팝업 관련 기능 -> terms.js
